refactor: move deletion modal file
