+++ src/actions/calendar-actions-fn.tsx
@@ -15,11 +15,11 @@ import { FOLDER_OPERATIONS } from '../constants/api';
 import { getFolderRequest } from '../soap/get-folder-request';
 import { getShareInfoRequest } from '../soap/get-share-info-request';
 import { CreateGroupModal } from './modals/create-group-modal';
+import { DeleteModal } from './modals/delete-modal';
 import { folderAction } from '../store/actions/calendar-actions';
 import { StoreProvider } from '../store/redux';
 import { ActionsClick } from '../types/actions';
 import { NewModal } from '../view/move/new-calendar-modal';
-import { DeleteModal } from '../view/sidebar/delete-modal';
 import { EditModal } from '../view/sidebar/edit-modal/edit-modal';
 import { EmptyModal } from '../view/sidebar/empty-modal';
 import { ShareCalendarModal } from '../view/sidebar/share-calendar-modal';
